refactor(command): extract command functions to command lib
