+++ packages/command/jest.config.ts
@@ -1,0 +1,14 @@
+/* eslint-disable */
+export default {
+    displayName: '@/command',
+    preset: '../../jest.preset.js',
+    testEnvironment: 'node',
+    transform: {
+        '^.+\\.[tj]s$': [
+            'ts-jest',
+            { tsconfig: '<rootDir>/tsconfig.spec.json' },
+        ],
+    },
+    moduleFileExtensions: ['ts', 'js', 'html'],
+    coverageDirectory: '../../coverage/packages/command',
+};
