+++ packages/command/package.json
@@ -1,0 +1,11 @@
+{
+    "name": "@/command",
+    "version": "0.0.1",
+    "dependencies": {
+        "tslib": "^2.3.0",
+        "@nx/devkit": "17.0.0"
+    },
+    "type": "commonjs",
+    "main": "./src/index.js",
+    "typings": "./src/index.d.ts"
+}
