+++ packages/command/src/index.ts
@@ -1,0 +1,6 @@
+export { runCommandFromDirectory } from './lib/runCommandFromDirectory/runCommandFromDirectory';
+export { executeCommandForFiles } from './lib/executeCommandForFiles/executeCommandForFiles';
+export { runCommand } from './lib/runCommand/runCommand';
+export { checkCommandExists } from './lib/checkCommandExists/checkCommandExists';
+export { commandExists } from './lib/commandExists/commandExists';
+export { executeCommand } from './lib/executeCommand/executeCommand';
