+++ packages/command/src/lib/checkCommandExists/checkCommandExists.spec.ts
