+++ packages/command/src/lib/checkCommandExists/checkCommandExists.ts
