+++ packages/command/src/lib/executeCommandForFiles/executeCommandForFiles.spec.ts
