+++ packages/command/src/lib/executeCommandForFiles/executeCommandForFiles.ts
