+++ packages/command/src/lib/runCommand/runCommand.spec.ts
