+++ packages/command/src/lib/runCommandFromDirectory/runCommandFromDirectory.ts
