+++ packages/util/src/index.ts
@@ -5,7 +5,6 @@ export { getProjectConfigurationType } from './lib/getProjectConfigurationType/g
 export { getProjectTypeFromConfig } from './lib/getProjectTypeFromConfig/getProjectTypeFromConfig';
 export { getProjectFromFile } from './lib/getProjectFromFile/getProjectFromFile';
 export { getWorkspaceLayout } from './lib/getWorkspaceLayout/getWorkspaceLayout';
-export { runCommandFromDirectory } from './lib/runCommandFromDirectory/runCommandFromDirectory';
 export { snakeCaseToCamelCase } from './lib/snakeCaseToCamelCase/snakeCaseToCamelCase';
 export { trimLib } from './lib/trimLib/trimLib';
 export { trimTest } from './lib/trimTest/trimTest';
@@ -17,11 +16,6 @@ export { getNxJsonConfiguration } from './lib/getNxJsonConfiguration/getNxJsonCo
 export { getLanguageExtension } from './lib/getLanguageExtension/getLanguageExtension';
 export { extractRootsFromExecutorContext } from './lib/extractRootsFromExecutorContext/extractRootsFromExecutorContext';
 export { formatNxJson } from './lib/formatNxJson/formatNxJson';
-export { executeCommandForFiles } from './lib/executeCommandForFiles/executeCommandForFiles';
-export { runCommand } from './lib/runCommand/runCommand';
-export { checkCommandExists } from './lib/checkCommandExists/checkCommandExists';
-export { commandExists } from './lib/commandExists/commandExists';
-export { executeCommand } from './lib/executeCommand/executeCommand';
 export { mergeArrays } from './lib/mergeArrays/mergeArrays';
 export { isDarwin } from './lib/isDarwin/isDarwin';
 export { isWindows } from './lib/isWindows/isWindows';
