+++ plugins/nx-cmake/.eslintrc.json
@@ -27,6 +27,7 @@
                             "@/util",
                             "@/mocks",
                             "@/types",
+                            "@/command",
                             "@/log"
                         ],
                         "includeTransitiveDependencies": true
