+++ plugins/nx-cmake/src/executors/cmake/configureProjectWithCMake/configureProjectWithCMake.spec.ts
@@ -1,9 +1,9 @@
 import { configureProjectWithCMake } from './configureProjectWithCMake';
-import * as getCmakeCommandArgumentsModule from '../getCmakeCommandArguments/getCmakeCommandArguments';
-import * as utilsModule from '@/util';
-import * as getCmakeModule from '../../../config/getPrograms/getCmake/getCmake';
 import { CmakeExecutorSchema } from '../../executor';
 import { LINUX_CMAKE } from '../../../config/getPrograms/getPrograms';
+import * as getCmakeCommandArgumentsModule from '../getCmakeCommandArguments/getCmakeCommandArguments';
+import * as getCmakeModule from '../../../config/getPrograms/getCmake/getCmake';
+import * as commandModule from '@/command';
 
 describe('buildProjectWithMake', () => {
     let workspaceRoot: string;
@@ -20,7 +20,7 @@ describe('buildProjectWithMake', () => {
             args: [],
             release: false,
         };
-        runCommandMock = jest.spyOn(utilsModule, 'runCommand');
+        runCommandMock = jest.spyOn(commandModule, 'runCommand');
         getCmakeCommandArgumentsMock = jest.spyOn(
             getCmakeCommandArgumentsModule,
             'getCmakeCommandArguments',
